lygtais viska sutvarkiau ko aiste prase...
