+++ src/main/java/org/coffee/service/BusinessService.java
@@ -35,30 +35,4 @@ public class BusinessService {
     public Long getActiveBusinessId() {
         return getActiveBusiness().getId();
     }
-
-    /**
-     * Change a user's password.
-     * @param userId          the database ID of the user
-     * @param currentPassword the existing plain‐text password to verify
-     * @param newPassword     the new plain‐text password
-     * @throws IllegalArgumentException if the current password is wrong or the user is not found
-     */
-    @Transactional
-    public void updatePassword(Long userId, String currentPassword, String newPassword) {
-        User user = userDAO.findById(userId);
-        if (user == null) {
-            throw new IllegalArgumentException("User not found");
-        }
-
-        // verify current password
-        if (!PasswordUtil.checkPassword(currentPassword, user.getPassword())) {
-            throw new IllegalArgumentException("Current password is incorrect");
-        }
-
-        // hash & set new password
-        user.setPassword(PasswordUtil.hashPassword(newPassword));
-
-        // merge back to the database
-        userDAO.merge(user);
-    }
 }
