+++ src/main/webapp/user/change-password.xhtml
@@ -1,35 +1,53 @@
 <!DOCTYPE html>
 <html xmlns="http://www.w3.org/1999/xhtml"
-      xmlns:h="http://xmlns.jcp.org/jsf/html">
+      xmlns:h="http://xmlns.jcp.org/jsf/html"
+      xmlns:p="http://primefaces.org/ui">
 <h:head>
     <title>Change Password</title>
 </h:head>
 <h:body>
-    <h:form>
-        <h:messages globalOnly="true" layout="table"/>
+    <h:form id="changePwdForm">
+        <!-- show global messages -->
+        <p:growl id="growl" showDetail="true" life="5000"/>
 
-        <h:panelGrid columns="2" columnClasses="label,value">
-            <h:outputLabel for="currentPwd" value="Current Password:"/>
-            <h:inputSecret id="currentPwd"
-                           value="#{changePasswordBean.currentPassword}"
-                           required="true"
-                           requiredMessage="Current password is required."/>
+        <p:panel header="Change Password">
+            <p:messages id="msgs" autoUpdate="true" />
 
-            <h:outputLabel for="newPwd" value="New Password:"/>
-            <h:inputSecret id="newPwd"
-                           value="#{changePasswordBean.newPassword}"
-                           required="true"
-                           requiredMessage="New password is required."/>
+            <p:panelGrid columns="2" columnClasses="label,value">
+                <!-- Current Password -->
+                <h:outputLabel for="currentPwd" value="Current Password:"/>
+                <p:password id="currentPwd"
+                            value="#{changePasswordBean.currentPassword}"
+                            required="true"
+                            requiredMessage="Current password is required."
+                            feedback="false"/>
+                <p:message for="currentPwd"/>
 
-            <h:outputLabel for="confirmPwd" value="Confirm Password:"/>
-            <h:inputSecret id="confirmPwd"
-                           value="#{changePasswordBean.confirmPassword}"
-                           required="true"
-                           requiredMessage="Please confirm new password."/>
-        </h:panelGrid>
+                <!-- New Password, with built-in confirm match -->
+                <h:outputLabel for="newPwd" value="New Password:"/>
+                <p:password id="newPwd"
+                            value="#{changePasswordBean.newPassword}"
+                            required="true"
+                            requiredMessage="New password is required."
+                            feedback="true"
+                            match="confirmPwd"/>
+                <p:message for="newPwd"/>
 
-        <h:commandButton value="Change Password"
-                         action="#{changePasswordBean.changePassword}"/>
+                <!-- Confirm Password -->
+                <h:outputLabel for="confirmPwd" value="Confirm Password:"/>
+                <p:password id="confirmPwd"
+                            value="#{changePasswordBean.confirmPassword}"
+                            required="true"
+                            requiredMessage="Please confirm new password."
+                            feedback="false"/>
+                <p:message for="confirmPwd"/>
+            </p:panelGrid>
+
+            <p:commandButton value="Change Password"
+                             action="#{changePasswordBean.changePassword}"
+                             update="@form"
+                             icon="pi pi-check"/>
+        </p:panel>
     </h:form>
 </h:body>
 </html>
